Fix for 32 bit builds and skip needless conversion
* Don't use 64 bit int in tests so 32 bit builds don't error. While this is kinda skipping a case, I believe this would have thrown an error anyways. Allowing bits of any width would require changing how constants in downstream computations are generated. It's probably only reasonable to have integers of less than or equal to system width here anyways.
* Also don't add a conversion if it's not needed (made conversion during unpacking of shape tuple conditional).

+++ numba/tests/test_dyn_array.py
@@ -621,7 +621,7 @@ class TestNdZeros(ConstructorBaseTest, TestCase):
         # Test for issue #4575
         pyfunc = self.pyfunc
         def func(m, n):
-            return pyfunc((np.int16(m), np.int64(n)))
+            return pyfunc((np.int16(m), np.int32(n)))
         self.check_2d(func)
 
     @tag('important')
@@ -696,7 +696,7 @@ class TestNdFull(ConstructorBaseTest, TestCase):
     def test_2d_shape_dtypes(self):
         # Test for issue #4575
         def func(m, n):
-            return np.full((np.int16(m), np.int64(n)), 4.5)
+            return np.full((np.int16(m), np.int32(n)), 4.5)
         self.check_2d(func)
